Add HasSaveLoadConfig to PlotsTableWidget
