+++ pyqtgraph_scope_plots/csv/csv_plots.py
@@ -209,32 +209,6 @@ class CsvLoaderPlotsTableWidget(AnimationPlotsTableWidget, PlotsTableWidget, Has
         )
         self._recents.bind_hotkeys(self)
 
-    @classmethod
-    def _get_all_model_bases(cls) -> List[Type[BaseModel]]:
-        bases = super()._get_all_model_bases()
-        plot_bases = cls._PLOT_TYPE._get_all_model_bases()
-        table_bases = cls._TABLE_TYPE._get_all_model_bases()
-        return bases + plot_bases + table_bases
-
-    @classmethod
-    def _get_data_model_bases(cls) -> List[Type[BaseModel]]:
-        bases = super()._get_data_model_bases()
-        plot_bases = cls._PLOT_TYPE._get_data_model_bases()
-        table_bases = cls._TABLE_TYPE._get_data_model_bases()
-        return bases + plot_bases + table_bases
-
-    def _write_model(self, model: BaseModel) -> None:
-        super()._write_model(model)
-        assert isinstance(self._table, HasSaveLoadDataConfig)
-        self._table._write_model(model)
-        self._plots._write_model(model)
-
-    def _load_model(self, model: BaseModel) -> None:
-        super()._load_model(model)
-        assert isinstance(self._table, HasSaveLoadDataConfig)
-        self._table._load_model(model)
-        self._plots._load_model(model)
-
     def _on_legend_checked(self) -> None:
         assert isinstance(self._plots, FullPlots)
         assert isinstance(self._table, FullSignalsTable)
